Deprecate wavefront Shift
This is a holdover from when I was thinking about the propagation core
a lot differently. In its current implementation, I can't come up with
any real world use of this object. It doesn't have any physical meaning
since any real shift would be introduced with a Tilt, so I'm just going
to call it dead code and nuke it.