Functionalize prop_mono
prop_mono is getting a little lengthy and it will only continue to grow
as more propagation functionality is built out. It will be better in the
long-run if this entire section of the code is functionalized.

While doing this work, it became clear that there wasn't any real
benefit to implementing the Fourier transforms as Wavefront class
methods. Wavefront has been simplified quite a bit and is mostly just a
data class at this point. It probably makes more sense to move the
Fourier transforms out of Wavefront and put the functionality in the
prop module instead.