More closely hit requested RMS
The power spectrum-based WFE method accepts an RMS attribute. While the
resulting phase map was getting somewhat close to the requested RMS, it
regularly coming in quite a bit lower than the requested value. This
commit does not guarantee that the requested RMS is hit exactly, but it
gets a lot closer than before.