Redefine internal cached_properties as standard properties
There is no real benefit in defining the internal cacheable
Plane attributes (amplitude, phase, and ptt_vector) as
cached_properties. It may in fact be slower, and it causes
the attributes to be misidentified in the documentation.

This commit modifies the Plane amplitude, phase, and ptt_vector
getters to be generic Python properties while maintaining the
cache behavior.

Resolves #7