Update _multiply_angle to use new Tilt definition
Previously Tilt was expressed in terms of image plane x and y but
it should really be expressed in terms of rotation about the x and
y axis of the Tilt plane. Commit 01da13b makes this change. This
commit updates the two usages of Tilt internal to Lentil - both
implementations of _multiply_angle().