Use functools.lru_cache to cache dft2 matrices
This commit builds on the caching work that Shanti did in c0bc6a7.
It employes a two-stage caching approach that will catch more cases
and employs lru_cache to ensure the caches won't grow out of
control.

* A cache on the computation of X, Y, U, and V will catch all cases
  where f.shape and npix are the same. This provides the primary
  performance gain.
* A cache on the computation of E1 and E2 will catch cases where
  shift and alpha are the same. This does in fact happen when
  performing propagations on multi-segmented models.
* A type check is performed on out to ensure we can cast complex
  data there.
* Updated the unit test to enable complex casting.

Overall this provides a 3x speedup on top of the performance gains
Shanti saw.